refactor: moveCar함수 getMovingCarsResults함수로 대체
이름으로 예측할 수 있는 한가지 일만 수행
